requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
tiingo==0.12.0
urllib3==1.25.9
coverage>=4.5.4
flake8==3.7.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
Flask-WTF==0.14.3
psycopg2-binary==2.8.4
Werkzeug==0.16.1
